Fix compiler warnings across all firmware targets (#42)
- Fix format specifier mismatches: cast uint32_t args to (unsigned int) for
  %u format in BUS_DEBUG calls, and HAL_UART_StateTypeDef to (int) for %d
- Remove unused variable last_time_us in motion task
- Suppress unused variable warning for posted in RX callback (used only
  when debug is enabled)

+++ firmware/shared/tsumikoro_bus/hal/stm32/tsumikoro_hal_stm32.c
@@ -303,7 +303,7 @@ tsumikoro_hal_status_t tsumikoro_hal_transmit(tsumikoro_hal_handle_t handle,
         }
 #ifdef TSUMIKORO_HAL_MOCK_DEBUG
         printf("[HAL_TX] ERROR: HAL_UART_Transmit_DMA failed, status=%d, UART gState=%d\n",
-               status, device->uart_handle.gState);
+               status, (int)device->uart_handle.gState);
 #endif
         return TSUMIKORO_HAL_ERROR;
     }
